Move four functions from `plumbing.rs` to `execution.rs`
These four functions are all only used within `execution.rs`, so they no
longer need `pub(crate)`.

This is the first step toward eliminating `plumbing.rs`.
