+++ compiler/rustc_query_impl/src/plumbing.rs
@@ -1,6 +1,3 @@
-use std::num::NonZero;
-
-use rustc_data_structures::Limit;
 use rustc_data_structures::unord::UnordMap;
 use rustc_middle::bug;
 #[expect(unused_imports, reason = "used by doc comments")]
@@ -8,72 +5,13 @@ use rustc_middle::dep_graph::DepKindVTable;
 use rustc_middle::dep_graph::{DepNode, DepNodeIndex, DepNodeKey, SerializedDepNodeIndex};
 use rustc_middle::query::erase::{Erasable, Erased};
 use rustc_middle::query::on_disk_cache::{CacheDecoder, CacheEncoder};
-use rustc_middle::query::{QueryCache, QueryJobId, QueryVTable, erase};
+use rustc_middle::query::{QueryCache, QueryVTable, erase};
 use rustc_middle::ty::TyCtxt;
-use rustc_middle::ty::tls::{self, ImplicitCtxt};
 use rustc_middle::verify_ich::incremental_verify_ich;
 use rustc_serialize::{Decodable, Encodable};
-use rustc_span::def_id::LOCAL_CRATE;
 
-use crate::diagnostics::{QueryOverflow, QueryOverflowNote};
 use crate::execution::{all_inactive, should_verify_loaded_value};
-use crate::job::find_dep_kind_root;
 use crate::query_vtables::for_each_query_vtable;
-use crate::{CollectActiveJobsKind, collect_active_query_jobs};
-
-fn depth_limit_error<'tcx>(tcx: TyCtxt<'tcx>, job: QueryJobId) {
-    let job_map = collect_active_query_jobs(tcx, CollectActiveJobsKind::Full);
-    let (span, desc, depth) = find_dep_kind_root(tcx, job, job_map);
-
-    let suggested_limit = match tcx.recursion_limit() {
-        Limit(0) => Limit(2),
-        limit => limit * 2,
-    };
-
-    tcx.dcx().emit_fatal(QueryOverflow {
-        span,
-        note: QueryOverflowNote { desc, depth },
-        suggested_limit,
-        crate_name: tcx.crate_name(LOCAL_CRATE),
-    });
-}
-
-#[inline]
-pub(crate) fn next_job_id<'tcx>(tcx: TyCtxt<'tcx>) -> QueryJobId {
-    QueryJobId(
-        NonZero::new(tcx.query_system.jobs.fetch_add(1, std::sync::atomic::Ordering::Relaxed))
-            .unwrap(),
-    )
-}
-
-#[inline]
-pub(crate) fn current_query_job() -> Option<QueryJobId> {
-    tls::with_context(|icx| icx.query)
-}
-
-/// Executes a job by changing the `ImplicitCtxt` to point to the new query job while it executes.
-#[inline(always)]
-pub(crate) fn start_query<R>(
-    job_id: QueryJobId,
-    depth_limit: bool,
-    compute: impl FnOnce() -> R,
-) -> R {
-    tls::with_context(move |icx| {
-        if depth_limit && !icx.tcx.recursion_limit().value_within_limit(icx.query_depth) {
-            depth_limit_error(icx.tcx, job_id);
-        }
-
-        // Update the `ImplicitCtxt` to point to our new query job.
-        let icx = ImplicitCtxt {
-            query: Some(job_id),
-            query_depth: icx.query_depth + if depth_limit { 1 } else { 0 },
-            ..*icx
-        };
-
-        // Use the `ImplicitCtxt` while we execute the query.
-        tls::enter_context(&icx, compute)
-    })
-}
 
 pub(crate) fn encode_query_values<'tcx>(tcx: TyCtxt<'tcx>, encoder: &mut CacheEncoder<'_, 'tcx>) {
     for_each_query_vtable!(CACHE_ON_DISK, tcx, |query| {
